HID: wacom: remove warning while disconnecting devices
When the LED class gets removed, it actually tries to reset the LED.
However, the device being disconnected, the set_report fails.

Previously, the attempt to cut lose this last event was through unsetting
the HID drvdata, but it was not working properly. Simply reset the LED
groups to NULL makes a more efficient solution.

Signed-off-by: Benjamin Tissoires <benjamin.tissoires@redhat.com>
Acked-by: Jason Gerecke <jason.gerecke@wacom.com>
Signed-off-by: Jiri Kosina <jkosina@suse.cz>
[jason.gerecke@wacom.com: Imported into input-wacom repository (c0265a9)]
Signed-off-by: Jason Gerecke <jason.gerecke@wacom.com>
[aaron.skomra@wacom.com: Backported from input-wacom repository (fa2a43d)]
Signed-off-by: Aaron Armstrong Skomra <aaron.skomra@wacom.com>