WIP: Improve HID parsing
The lower two bits of a HID item specifies the length of the data
that follows. Instead of using hard-coded lengths to increment the
loop counter 'i', read the actual length. This allows us to properly
move through more complex descriptors that may have some items whose
data is of an unexpected length. Additionally, with the length known,
we can extract out the complete "data" and use it instead of accessing
the report bytes as-needed.