2.6.30: Fix maximum distance for Bamboo tablets
Way back when, the driver used to report the distance data from Bamboo
tablets unmodified. The unmodified data *increases* as you get closer
to the tablet, however, which is the opposite of what is expected. Chris
Bagwell changed this behavior in 2011 with upstream commit c18c2cec31.
At the same time, he changed the maximum distance from 63 to 31. Although
Chris eventually backported the former change here to the 2.6.30 tree as
part of commit e9ddd7f10b, the latter change was never made. The 2.6.38
and higher trees already have this change.

Fixes: e9ddd7f10b ("Bamboo 2nd and 3rd gen support for 2.6.30+ kernels")
Signed-off-by: Jason Gerecke <jason.gerecke@wacom.com>
Reviewed-by: Ping Cheng <ping.cheng@wacom.com>