Bump minimum required kernel from 2.6.30 to 2.6.32
The xf86-input-wacom driver sets the resolution of X and Y axes so that
clients can relate logical units to physical distances. This information
is normally provided by the kernel, with the exception of when the 2.6.30
tree is in use. The kernel only gained the ability to report resolution
in the 2.6.31 kernel, requring xf86-input-wacom to keep around a parallel
table specifically for users who have a 2.6.30. Since RHEL 6 is the last
remaining user of this tree, it is safe to bump up the required input-wacom
version up 2.6.32.

Signed-off-by: Jason Gerecke <jason.gerecke@wacom.com>