2.6.32: Backport resolution support to 2.6.32
input_absinfo did not have resolution information before 2.6.32. We
used a non-standard method (stuffing the physical X and Y maximum
into ABS_RX and ABS_RY) in 2.6.30 tree to support resolution for pen
and touch hybrid devices. Now that we've moved the minimum version of
this tree to 2.6.32, we're able to set resolution to input_absinfo.
However, resolution was not reported to userland before 2.6.36. We
need to backport input_abs_set_res to 2.6.32 tree. But, RHEL 6 did
that for their kernels. To avoid the compile error on RHEL 6, we use
wacom_input_abs_set_res in place of input_abs_set_res.

Signed-off-by: Jason Gerecke <jason.gerecke@wacom.com>
Signed-off-by: Ping Cheng <ping.cheng@wacom.com>