2.6.32: Fix abs get/set accessors for 2.6.36/2.6.37
The 2.6.36 kernel removes several members from `struct input_dev` (e.g.
abs, absres) and replaces them with a structs and accessor functions.
To allow the input-wacom code to compile under both old and new kernels,
commits ca9786f and the mailinglist version of ab2ea683fb conditionally
defined their own implementation of the accessor function for older
kernels.

It was noticed, however, that this did not compile correctly on RHEL 6
systems. It seems that the accessor API introduced in 2.6.36 is provided
in their customized "2.6.32" kernel. This results in a redefinition error
that halts compilation. To work around this, commit 58d8320541 removed
the condtional and renamed our implementation of the accessor. Commit
ab2ea683fb was also similarly modified from its mailinglist version prior
to being committed. This change prevented the redefinition on RHEL 6 and
also worked fine for pre-2.6.36 kernels. The change ended up breaking
compilation under stock 2.6.36 kernels since the members used by the
renamed function were removed.

To ensure the code compiles in all cases, we need to be a little more
clever. We make use of the recently-added "WACOM_LINUX_TRY_COMPILE"
configure macro to see if the kernel provides the accessor API or not.
If it does, we make use of it; of not, we access the members directly.

Fixes: 58d8320541 ("2.6.30: define wacom_input_abs_get_val")
Fixes: ab2ea683fb ("2.6.32: Backport resolution support to 2.6.32")
Signed-off-by: Jason Gerecke <jason.gerecke@wacom.com>
Reviewed-by: Ping Cheng <ping.cheng@wacom.com>