HID: wacom: generic: read the contact count max from reports other than 0x23
All previous contact count maxes in the generic code path were
read from the second byte of report 0x23.

In this specific case, report 0x82 has the lenght in the higher
nibble of the third byte. We should support reading the
contact count max no matter what report we are reading or
which position the max is in.

To do this we submit the whole feature report as a report
using hid_report_raw_event. Our event path catches
the contact count max when it sees the HID_DG_CONTACTMAX
usage.

Signed-off-by: Aaron Armstrong Skomra <aaron.skomra@wacom.com>