travis: Scan every commit for style issues
The HID and input maintainers are becoming more strict about the use of
checkpatch.pl prior to patch submission. Since development often takes
place in the input-wacom tree first and Travis already checks for various
issues, we might as well have it also perform the style checks.

Because we want to scan each commit (there could be issues with the e.g.
commit description) we write patches for every commit in this branch
which is not already on master and run checkpatch on the lot of them.

Signed-off-by: Jason Gerecke <jason.gerecke@wacom.com>