travis: Properly handle commits pushed/merged to master
Pushing or merging a branch to master updates the repository's reference
to that branch. When our travis script fetches the master branch and calls
`git format-patch` to produce a list of commits that aren't already on
master, there is obviously nothing to generate. This isn't a problem on
its own, but checkpatch the errors out with:

scripts/checkpatch.pl: *.patch: open failed - No such file or directory

To work around this we now only run checkpatch if patch files were
generated. Because we should only be pushing commits that were already
tested, this shouldn't significantly change our coverage.

Signed-off-by: Jason Gerecke <jason.gerecke@wacom.com>